# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CodeArena API",
    "description": "Production API specification for CodeArena — Online Judge & Secure Code Execution Platform.",
    "version": "1.0.0",
    "contact": {
      "name": "CodeArena Engineering"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Environment Server"
    }
  ],
...
  "paths": {
...
    "/api/v1/problems": {
      "get": {
        "tags": ["Problems"],
        "summary": "List active problems",
        "description": "Public endpoint to retrieve active problems with summary projections, test case counts, readiness breakdowns, pagination, and derived solved status for authenticated users.",
        "operationId": "listProblems",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (minimum 1)",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of problems per page (1 to 100)",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "difficulty",
            "in": "query",
            "description": "Filter by difficulty tier",
            "schema": {
              "type": "string",
              "enum": ["EASY", "MEDIUM", "HARD"]
            }
          },
          {
            "name": "tag",
            "in": "query",
            "description": "Filter by single topic tag (alias for tags)",
            "schema": {
              "type": "string",
              "example": "array"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "description": "Filter by topic tag(s)",
            "schema": {
              "type": "string",
              "example": "array,math"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Case-insensitive title search query",
            "schema": {
              "type": "string",
              "example": "Sum"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria for problem list",
            "schema": {
              "type": "string",
              "enum": ["newest", "oldest", "title_asc", "title_desc"],
              "default": "newest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of active problems",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "problems": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ProblemSummary"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Problems"],
        "summary": "Create a problem",
        "description": "Creates a new problem. Restricted strictly to administrators (ADMIN role). Records administrative audit log.",
        "operationId": "createProblem",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProblemRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Problem created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "problem": {
                          "$ref": "#/components/schemas/Problem"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/api/v1/problems/{problemId}": {
      "get": {
        "tags": ["Problems"],
        "summary": "Get problem details by ID",
        "description": "Public endpoint to retrieve complete specification and examples for an active problem.",
        "operationId": "getProblemById",
        "parameters": [
          {
            "name": "problemId",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the problem",
            "schema": {
              "type": "string",
              "example": "65f0123456789abcdef01235"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Problem details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "problem": {
                          "$ref": "#/components/schemas/Problem"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "patch": {
        "tags": ["Problems"],
        "summary": "Update an existing problem",
        "description": "Partially updates an existing problem. Restricted strictly to administrators (ADMIN role). Records administrative audit log.",
        "operationId": "updateProblem",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "problemId",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the problem to update",
            "schema": {
              "type": "string",
              "example": "65f0123456789abcdef01235"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProblemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Problem updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "problem": {
                          "$ref": "#/components/schemas/Problem"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "delete": {
        "tags": ["Problems"],
        "summary": "Deactivate a problem",
        "description": "Soft-deactivates an existing problem (`isActive: false`). Restricted strictly to administrators (ADMIN role). Records administrative audit log.",
        "operationId": "deactivateProblem",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "problemId",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the problem to deactivate",
            "schema": {
              "type": "string",
              "example": "65f0123456789abcdef01235"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Problem deactivated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Problem deactivated successfully" }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
    "/api/v1/admin/audit-logs": {
      "get": {
        "tags": ["Admin"],
        "summary": "List administrative audit logs",
        "description": "Retrieves audit trails for administrative problem and test-case mutations. Restricted strictly to administrators (ADMIN role).",
        "operationId": "getAuditLogs",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (minimum 1)",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of logs per page (1 to 100)",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "action",
            "in": "query",
            "description": "Filter by audit action type",
            "schema": {
              "type": "string",
              "enum": [
                "PROBLEM_CREATED",
                "PROBLEM_UPDATED",
                "PROBLEM_DELETED",
                "TEST_CASE_CREATED",
                "TEST_CASE_UPDATED",
                "TEST_CASE_DELETED"
              ]
            }
          },
          {
            "name": "targetType",
            "in": "query",
            "description": "Filter by target entity type",
            "schema": {
              "type": "string",
              "enum": ["Problem", "TestCase"]
            }
          },
          {
            "name": "targetId",
            "in": "query",
            "description": "Filter by target entity ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "Filter by originating user ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated audit logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "logs": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/AuditLog"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Provide a valid JWT token issued by POST /api/v1/auth/login"
      }
    },
    "responses": {
      "BadRequestError": {
        "description": "Bad Request — Invalid parameter, malformed payload, or validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Title is required and must not be empty",
              "errorCode": "VALIDATION_ERROR",
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "Title is required and must not be empty"
              },
              "requestId": "req_mu576kgd_5826a9e8"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Unauthorized — Missing, invalid, or expired authentication token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Invalid authorization header format. Expected Bearer <token>",
              "errorCode": "AUTHENTICATION_REQUIRED",
              "error": {
                "code": "AUTHENTICATION_REQUIRED",
                "message": "Invalid authorization header format. Expected Bearer <token>"
              },
              "requestId": "req_mu576kgd_5826a9e8"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Forbidden — Insufficient permissions or object-level authorization denial",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Access denied: insufficient permissions",
              "errorCode": "FORBIDDEN",
              "error": {
                "code": "FORBIDDEN",
                "message": "Access denied: insufficient permissions"
              },
              "requestId": "req_mu576kgd_5826a9e8"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Not Found — Requested resource does not exist or is inactive",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Problem not found",
              "errorCode": "RESOURCE_NOT_FOUND",
              "error": {
                "code": "RESOURCE_NOT_FOUND",
                "message": "Problem not found"
              },
              "requestId": "req_mu576kgd_5826a9e8"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ProblemExample": {
        "type": "object",
        "required": ["input", "output"],
        "properties": {
          "input": {
            "type": "string",
            "example": "2\n1 2\n"
          },
          "output": {
            "type": "string",
            "example": "3\n"
          }
        }
      },
      "ProblemReadiness": {
        "type": "object",
        "properties": {
          "isReady": {
            "type": "boolean",
            "example": true
          },
          "publicCount": {
            "type": "integer",
            "example": 2
          },
          "hiddenCount": {
            "type": "integer",
            "example": 3
          },
          "totalCount": {
            "type": "integer",
            "example": 5
          },
          "missingRequirements": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": []
          }
        }
      },
      "ProblemSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "65f0123456789abcdef01235"
          },
          "title": {
            "type": "string",
            "example": "Sum of Two Numbers"
          },
          "difficulty": {
            "type": "string",
            "enum": ["EASY", "MEDIUM", "HARD"],
            "example": "EASY"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["math", "basics"]
          },
          "authorId": {
            "type": "string",
            "example": "65f0123456789abcdef01234"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "solved": {
            "type": "boolean",
            "description": "Indicates whether the authenticated user has solved this problem with an ACCEPTED verdict",
            "example": false
          },
          "testCasesCount": {
            "type": "integer",
            "example": 5
          },
          "readiness": {
            "$ref": "#/components/schemas/ProblemReadiness"
          }
        }
      },
      "Problem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "65f0123456789abcdef01235"
          },
          "title": {
            "type": "string",
            "example": "Sum of Two Numbers"
          },
          "description": {
            "type": "string",
            "example": "Given two integers A and B, calculate and return their sum."
          },
          "difficulty": {
            "type": "string",
            "enum": ["EASY", "MEDIUM", "HARD"],
            "example": "EASY"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["math", "basics"]
          },
          "inputFormat": {
            "type": "string",
            "example": "The first line contains an integer T. Each test case consists of two integers A and B."
          },
          "outputFormat": {
            "type": "string",
            "example": "For each testcase, output the sum A + B on a new line."
          },
          "constraints": {
            "type": "string",
            "example": "1 <= T <= 100\n-10^9 <= A, B <= 10^9"
          },
          "examples": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemExample"
            }
          },
          "authorId": {
            "type": "string",
            "example": "65f0123456789abcdef01234"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "testCasesCount": {
            "type": "integer",
            "example": 5
          },
          "readiness": {
            "$ref": "#/components/schemas/ProblemReadiness"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateProblemRequest": {
        "type": "object",
        "required": [
          "title",
          "description",
          "difficulty",
          "inputFormat",
          "outputFormat",
          "constraints",
          "examples"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "Sum of Two Numbers"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "example": "Given two integers A and B, return their sum."
          },
          "difficulty": {
            "type": "string",
            "enum": ["EASY", "MEDIUM", "HARD"],
            "example": "EASY"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["math", "basics"]
          },
          "inputFormat": {
            "type": "string",
            "example": "First line contains T. Each of the following T lines contains two integers A and B."
          },
          "outputFormat": {
            "type": "string",
            "example": "For each testcase, print the sum A + B on a new line."
          },
          "constraints": {
            "type": "string",
            "example": "1 <= T <= 100\n-10^9 <= A, B <= 10^9"
          },
          "examples": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/ProblemExample"
            }
          }
        }
      },
      "UpdateProblemRequest": {
        "type": "object",
        "description": "At least one field must be provided for update.",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "Updated Sum of Two Numbers"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "example": "Updated problem statement..."
          },
          "difficulty": {
            "type": "string",
            "enum": ["EASY", "MEDIUM", "HARD"],
            "example": "EASY"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["math"]
          },
          "inputFormat": {
            "type": "string",
            "example": "Two integers on a single line."
          },
          "outputFormat": {
            "type": "string",
            "example": "Single integer sum."
          },
          "constraints": {
            "type": "string",
            "example": "-10^9 <= A, B <= 10^9"
          },
          "examples": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/ProblemExample"
            }
          }
        }
      },
...
      "AuditLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "65f0123456789abcdef01238"
          },
          "userId": {
            "type": "string",
            "example": "65f0123456789abcdef01234"
          },
          "user": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": { "type": "string", "example": "65f0123456789abcdef01234" },
              "name": { "type": "string", "example": "Admin User" },
              "email": { "type": "string", "example": "admin@codearena.dev" }
            }
          },
          "action": {
            "type": "string",
            "enum": [
              "PROBLEM_CREATED",
              "PROBLEM_UPDATED",
              "PROBLEM_DELETED",
              "TEST_CASE_CREATED",
              "TEST_CASE_UPDATED",
              "TEST_CASE_DELETED"
            ],
            "example": "PROBLEM_CREATED"
          },
          "targetType": {
            "type": "string",
            "enum": ["Problem", "TestCase"],
            "example": "Problem"
          },
          "targetId": {
            "type": "string",
            "example": "65f0123456789abcdef01235"
          },
          "details": {
            "type": "object",
            "example": {
              "title": "Sum of Two Numbers",
              "difficulty": "EASY"
            }
          },
          "ip": {
            "type": "string",
            "nullable": true,
            "example": "127.0.0.1"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "total": {
            "type": "integer",
            "example": 42
          },
          "totalPages": {
            "type": "integer",
            "example": 3
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Resource not found"
          },
          "errorCode": {
            "type": "string",
            "example": "RESOURCE_NOT_FOUND"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "RESOURCE_NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Resource not found"
              }
            }
          },
          "requestId": {
            "type": "string",
            "example": "req_mu576kgd_5826a9e8"
          }
        }
      }
    }
  }
}